* lib/gnupload (GPG): Don't use an absolute path.
This reverts part of the 2004-01-28 change.
If the hypothetical cracker ever gets in to my (or any developer's)
system with sufficient privilege to modify the contents of
directories in my PATH (or change my PATH altogether), they can
already compromise my development work in so many ways that using
such absolute names in gnupload gives reduced functionality with
no added security.