Use `Exit' instead of `exit' in test suite.
Cater to Bourne shells like Solaris sh that do not pass the
`exit' argument as status to the cleanup trap.
* Makefile.am (maintainer-check): Check that here-documents
use only `END' or `EOF' as delimiter in the test suite.
Check that, outside of here-documents, the tests do not use
`exit' with an argument, but use `Exit' instead.
* tests/defs.in (Exit): New function.  Use it throughout,
starting with the introduction of the exit trap.
* tests/*.test: Use `Exit $arg' instead of `exit $arg'
throughout, except inside created files.

Signed-off-by: Ralf Wildenhues <Ralf.Wildenhues@gmx.de>