Do not reorder nobase ltlib installation.
This patch changes back the order of installation of nobase
LTLIBRARIES to that in which files are listed in the variables.
It does not fix the issue of strict dependency ordering, as
implied by *_LIBADD, but as before allows to steer the order by
listing the libraries in nobase_*_LTLIBRARIES appropriately.

As an advantage over the state of 1.10, multiple consecutive
nobase libraries to be installed in the same directory are
passed to libtool in one invocation.

* lib/am/ltlib.am (install-%DIR%LTLIBRARIES) [!%?BASE%]: Do not
reorder libraries upon installation.  Remove spurious empty
line.
* tests/ltorder.test: New test.
* tests/Makefile.am: Update.

Signed-off-by: Ralf Wildenhues <Ralf.Wildenhues@gmx.de>