Correctly treat all assignments with bogus trailing comments.
Before this patch, automake would silently fail to diagnose and
to copy into the output those variable assignments which are
preceded by a comment, and end in backslash newline comment.

* automake.in (read_am_file): When determining whether an
escaped newline followed by a comment is an error, correctly use
the parser state, not the contents of a saved comment, which
could still be carried over from a comment before an assignment.
* NEWS: Update.
* tests/commen11.test: New test.
* tests/Makefile.am: Update.
Report by Karl Berry.

Signed-off-by: Ralf Wildenhues <Ralf.Wildenhues@gmx.de>