Let AM_PATH_PYTHON honor python's idea about the site directory.
* m4/python.m4 (AM_PATH_PYTHON): When computing pythondir and
pyexecdir, pass the expanded prefix resp. exec_prefix as `prefix'
to get_python_lib, so python can determine the name of the site
directory depending on the install location.  Afterwards, replace
the directory names with the unexpanded values of $PYTHON_PREFIX
resp. $PYTHON_EXEC_PREFIX again, to allow override according to
the documentation.  Fixes site directory computation for Debian
and Ubuntu (`dist-packages' for a prefix of `/usr' or `/usr/local',
`site-packages' elsewhere).
* NEWS, THANKS: Update.

Signed-off-by: Ralf Wildenhues <Ralf.Wildenhues@gmx.de>