Testsuite: do not use `chmod -R' when cleaning up.
* tests/Makefile.am (clean-local-check): Do not use `chmod -R' on
the test directories, as that may change or try to change the mode
of installed files: the test directory may contain symlinks to
ltmain.sh files from a Libtool installation, and Solaris `chmod -R'
touches symlink targets.  Instead, use the cleanup strategy used
in distdir.am (which is also employed in tests/defs.in).

Signed-off-by: Ralf Wildenhues <Ralf.Wildenhues@gmx.de>