Internationalization tests: prefer `make distdir' over `make dist'.
Instead of running `make dist', then uncompressing the created
tarball, and then grepping the uncompressed tar archive, we can
simply run `make distdir' and analyze the contents of the created
distibution directory.  This is both faster and cleaner.

* tests/pot-msgidbugs.test: Test updated to use `make distdir'
instead of `make dist' where possible.
* tests/pot-no-linguas.test: Likewise.
* tests/pot-override1.test: Likewise.
* tests/pot-xgettext1.test: Likewise.
* tests/pot-xgettext2.test: Likewise.
* tests/pot-xgettext3.test: Likewise.
* tests/pot-xgettext4.test: Likewise.
* tests/pot-download1.test: Likewise, and check the created
distdir instead of the distribution tarball (which is not even
created anymore).
($distdir): New variable.
* tests/pot-format-csharp1.test: Likewise.
* tests/pot-format-java1.test: Likewise.
* tests/pot-format-java2.test: Likewise.
* tests/pot-format-qt1.test: Likewise.
* tests/pot-format-tcl1.test: Likewise.
* tests/pot-linguas.test: Likewise.
* tests/pot-noinst.test: Likewise.
* tests/pot-samedir.test: Likewise.
* tests/pot-topsrcdir.test: Likewise.