regex: deprecate the obsolete macro AM_WITH_REGEX
This is a backport of commit v1.11-433-g37b0aee.

Today, practically nobody uses the GNU rx library, which, according
to its own website <http://www.gnu.org/software/rx/rx.html>, has
been "decommissioned".  Consequently, the automake-provided macro
AM_WITH_REGEX is not used nor required anymore.  Deprecate it, so
that it will be possible to safely remove it in the next major
automake version.

* m4/regex.m4 (AM_WITH_REGEX): Give a warning of the class
`obsolete' when this macro is used.
* doc/automake.texi (Public Macros): Move description of
`AM_WITH_REGEX' from here ...
(Obsolete Macros): ... to here, and declare it as obsolete
and "to be removed in a future version".
* tests/regex-obsolete.test: New test.
* tests/Makefile.am (TESTS): Add it.
* NEWS: Update.

See also:
<http://lists.gnu.org/archive/html/automake-patches/2010-11/msg00131.html>