Fix regression in ansi.test (failure on AIX 5.3).
* tests/ansi.test: Remove redundant hackish check done using a
hand-postprocessed Makefile.in.  This checs worked by setting
the `$(SHELL)' macro on the command line of make, but this is
not supported by the AIX 5.3 make implementation.

This bug has been lurking for a long time, and was activated by
commit v1.11-125-gc1f6cdb "Enable `errexit' shell flag in various
tests".  Report by Ralf Wildenhues.