ylwrap: make less general and more "expert"
Related to automake bug#7648 and PR automake/491.

Using the fact that ylwrap is called only in specific ways from
the Automake-generated Makefiles, we rewrote parts of it to let
it have and use more knowledge of the building process.
This decreases the generality of the script, but increases its
clarity and, more important for us, the possibility for future
optimizations and extensions.

* lib/ylwrap: Partial rewrite to take advantage of the knowledge
about the possible ways this script is called, and the fact that
it deals only with Yacc/Lex and not with generic preprocessors
or generators of C source and header files.  Also improved error
messages and checks about invalid usage.
* lib/am/yacc.am (?GENERIC?%EXT%%DERIVED-EXT%, ?!GENERIC?%OBJ%):
Adjust to the new simplified ylwrap calling convention.
(am__yacc_c2h): Remove, it's not needed anymore.
* lib/am/yacc.am (?GENERIC?%EXT%%DERIVED-EXT%, ?!GENERIC?%OBJ%):
Adjust to the new simplified ylwrap calling convention.
* automake.in (lang_yacc_target_hook): Adjust comment.
* tests/yacclex-line.test: New test.
* tests/ylwrap-extra-args.test: Likewise.
* tests/ylwrap-fail.test: Likewise.
* tests/ylwrap-lex-parallel.test: Likewise.
* tests/ylwrap-output-rename.test: Likewise.
* tests/ylwrap-path-handling.test: Likewise.
* tests/ylwrap-signals.test: Likewise.
* tests/ylwrap-yacc-dos.test: Likewise.
* tests/ylwrap-yacc-ext.test: Likewise.
* tests/ylwrap-yacc-parallel.test: Likewise.
* tests/Makefile.am (TESTS): Update.