ylwrap: various refactorings, few improvements
* lib/ylwrap: Few more safety measures: abort if we fail to get the
current working directory or to chdir to the temporary directory,
and avoid doing the final cleanup if we fail to chdir back to the
original working directory.
Remove various useless use of quotes (e.g., in `case' statements
or in variable assignments).
Avoid using a runtime-determined fancy name for the temporary file
where to save the post-processed yacc-generated header; this saves
a couple of forks.
(quote_for_sed, dirname_, dos_fix_yacc_filenames): New helper
functions, to reduce code inlined in the main loop of the script.
($ocwd): New variable, contains the absolute path of the working
directory where the script was started.  This will avoid us the
need to call `pwd` multiple times.
($do_exit): Do not discard anymore the stdout/stderr of the final
`rm -rf' on the temporary directory; it's better let the user know
whether the cleanup failed.
* tests/ylwrap-pwd-fail.test: New test.
* tests/ylwrap-chdir-back-fail.test: New test.
* tests/ylwrap-cleanup-fail.test: New test.
* tests/Makefile.am (TESTS): Update.