ylwrap: get rid of the loop on "FROM TARGET" couples
* lib/ylwrap: The new shape in which the previous refactoring have
left the code allows us to get rid on the confusing loop on the
"FROM TARGET" couples, where FROM is the name of an output file
generated by the wrapped program and TARGET is the name of the
file it should be renamed to.  The renamings can now be done in
a more ad-hoc, conscious way.  This is also meant will allow
further optimizations and cleanups in the future.