lex tests: remove erroneous check about ylwrap distribution
* tests/lex5.test: Do not check that the ylwrap script is *not*
distributed when there is only one lexer, as ylwrap is in fact
expected to distributed unconditionally (at least starting from
commit `Release-1-9-323-gc5881cc' of 19-08-2006).  Note that this
bogus check wasn't triggering any failure because it was done at
the wrong time (i.e. from the VPATH build directory), so that it
could never fail anyway.
Instead, check that ylwrap is distributed even when there is only
one lexer.
Since we are at it, add a couple of `ls -l' calls, to get better
debugging info.