tests: fix typo-related error in auxdir2.test
* tests/auxdir2.test (configure.in):  Close m4 quoting in the
argument to AC_CONFIG_AUX_DIR.  Without this, aclocal fails with
"ERROR: end of file in string".  This problem hasn't been exposed
by the testsuite before because this test is in XFAIL_TESTS, so
its failure went unnoticed, even if it was due to a wrong cause.

Bug introduced in commit v1.11-249-g49ac3de.