Revert "docs: parallel-tests is not experimental anymore"
This reverts commit a9eef973b5ea47cc3495f1a8307d4f7b85aea46f.

It turned out that the current work to introduce TAP and SubUnit
support in Automake-generated testsuite harnesses will probably
require the introduction of slight incompatibilities in the
'parallel-tests' behaviour, starting from release 1.12 onward.
So it's advisable to continue to characterize the 'parallel-tests'
support as "experimental" in maintenance release 1.11.2.

Suggestion from Ralf Wildenhues.