lex tests: fix spurious link errors on Solaris
On Solaris 10, linking of lex-generated programs was failing in a
couple of lex tests with errors like this:
  g++ -g -O2 -o joe joe.o -ll
  Undefined symbol     first referenced in file
  yywrap()             joe.o
  ld: fatal: Symbol referencing errors. No output written to joe
  collect2: ld returned 1 exit status
This change fixes it, also fixing at once potential problems for
systems that don't have a "lex library" (this happens for example
when cross-compiling from GNU/Linux to MinGW).

* tests/lex-depend-cxx.test (joe.ll, moe.l++): Define a dummy
`yywrap()' function.
* tests/lex-clean-cxx.test (mainfoo.cc, mainbar.cpp, mainbaz.c++,
mainqux.cxx): Likewise.